# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/example/mvvm/activity/MainActivity.kt
#	app/src/main/java/com/example/mvvm/activity/ScreenActivity.kt
#	app/src/main/java/com/example/mvvm/activity/Test2Activity.kt
#	app/src/main/java/com/example/mvvm/activity/TestActivity.kt
#	app/src/main/java/com/example/mvvm/utils/FunctionsLive.kt
#	app/src/main/java/com/example/mvvm/utils/GestureManager.kt
#	app/src/main/java/com/example/mvvm/viewmodel/TestViewModel.kt
#	app/src/main/java/com/example/mvvm/widget/ArcProgress.kt
#	app/src/main/java/com/example/mvvm/widget/CodeInput.kt
#	app/src/main/java/com/example/mvvm/widget/automatic/holder/NormalPicHolder.kt
#	app/src/main/res/drawable/shape_code.xml
#	app/src/main/res/drawable/shape_code_pressed.xml
#	app/src/main/res/drawable/shape_r5_blue.xml
#	app/src/main/res/layout/activity_screen.xml
#	app/src/main/res/values/attrs.xml
#	lib_common/src/main/java/com/example/common/config/RouterPath.kt

+++ lib_common/src/main/java/com/example/common/base/BaseActivity.kt
@@ -50,7 +50,7 @@ import com.example.common.network.socket.topic.WebSocketObserver
 import com.example.common.utils.DataBooleanCache
 import com.example.common.utils.ScreenUtil.screenHeight
 import com.example.common.utils.ScreenUtil.screenWidth
-import com.example.common.utils.builder.shortToast
+import com.example.common.utils.builder.toast
 import com.example.common.utils.function.registerResultWrapper
 import com.example.common.utils.manager.AppManager
 import com.example.common.utils.permission.PermissionHelper
@@ -61,7 +61,7 @@ import com.example.common.utils.setStatusBarLightMode
 import com.example.common.widget.dialog.AppDialog
 import com.example.common.widget.dialog.LoadingDialog
 import com.example.common.widget.textview.edittext.SpecialEditText
-import com.example.framework.utils.builder.TimerBuilder
+import com.example.framework.utils.builder.TimerBuilder.Companion.schedule
 import com.example.framework.utils.function.color
 import com.example.framework.utils.function.getIntent
 import com.example.framework.utils.function.value.hasAnnotation
@@ -218,15 +218,17 @@ abstract class BaseActivity<VDB : ViewDataBinding> : AppCompatActivity(), BaseIm
         // 先检测大屏设备，避免布局加载
         if (checkLargeScreen()) {
             launch {
-                delay(800)
-                // 关闭所有Activity
-                finishAffinity()
-                // 终止进程（兼容所有安卓版本，捕获异常）
-                try {
-                    killProcess(myPid())
-                    exitProcess(0)
-                } catch (e: Exception) {
-                    e.printStackTrace()
+                delay(800L)
+                if (!isFinishing && !isDestroyed) {
+                    // 关闭所有Activity
+                    finishAffinity()
+                    // 终止进程（兼容所有安卓版本，捕获异常）
+                    try {
+                        killProcess(myPid())
+                        exitProcess(0)
+                    } catch (e: Exception) {
+                        e.printStackTrace()
+                    }
                 }
             }
             // 如果检测到大屏设备，直接return，不执行后续逻辑
@@ -235,6 +237,11 @@ abstract class BaseActivity<VDB : ViewDataBinding> : AppCompatActivity(), BaseIm
         initBefore()
         if (needTransparentOwner) {
             overridePendingTransition(R.anim.set_alpha_in, R.anim.set_alpha_none)
+            /**
+             * 在 Android 8.0 (API 26) 中，Google 引入了一个非常严格的限制：如果 Activity 是透明的（translucent）或浮动的（floating），则不允许通过代码或 Manifest 指定屏幕方向
+             * 一旦违反，系统会在 onCreate → setRequestedOrientation 时直接抛出：java.lang.IllegalStateException: Only fullscreen opaque activities can request orientation
+             * API 27+已修复，透明 Activity 又可以安全地设置 SCREEN_ORIENTATION_PORTRAIT 了
+             */
             requestedOrientation = if (Build.VERSION.SDK_INT == Build.VERSION_CODES.O) {
                 ActivityInfo.SCREEN_ORIENTATION_UNSPECIFIED
             } else {
@@ -264,10 +271,7 @@ abstract class BaseActivity<VDB : ViewDataBinding> : AppCompatActivity(), BaseIm
      * 检测大屏设备
      * @return true-检测到大屏设备并弹出提示，false-正常设备
      */
-    private var checkedLargeScreen = false
     private fun checkLargeScreen(): Boolean {
-        if (checkedLargeScreen) return false
-        checkedLargeScreen = true
         // 页面销毁直接返回
         if (isFinishing || isDestroyed) return false
         // 判断是否为大屏设备（宽度≥600dp）
@@ -276,7 +280,7 @@ abstract class BaseActivity<VDB : ViewDataBinding> : AppCompatActivity(), BaseIm
         val isPhysicalTablet = config.smallestScreenWidthDp >= 600
         // 只要是物理平板 → 直接拦截，不管是不是分屏
         if (isPhysicalTablet) {
-            "当前设备为平板/大屏设备，暂不支持使用".shortToast()
+            "当前设备为平板/大屏设备，暂不支持使用".toast()
         }
         return isPhysicalTablet
     }
@@ -316,55 +320,55 @@ abstract class BaseActivity<VDB : ViewDataBinding> : AppCompatActivity(), BaseIm
     }
 
     /**
-     * 1.如果当前设备支持状态栏字体变色，会设置状态栏字体为黑色/如果当前设备不支持状态栏字体变色，会使当前状态栏加上透明度，否则不执行透明度
-     * 2.导航栏字体深色或亮色，只支持android O(api26)以上版本,背景在5.0+可设置,为兼顾最低6.0(23+)的手机,统一底部为白色
-     *
+     * 1) 如果当前设备支持状态栏字体变色，会设置状态栏字体为黑色/如果当前设备不支持状态栏字体变色，会使当前状态栏加上透明度，否则不执行透明度
+     * 2) 导航栏字体深色或亮色，只支持 android O(api26) 以上版本,背景在 5.0+ 可设置,为兼顾最低 6.0(23+) 的手机,统一底部为白色
      * ImmersionBar.with(this)
-     *              .transparentStatusBar()  //透明状态栏，不写默认透明色
-     *              .transparentNavigationBar()  //透明导航栏，不写默认黑色(设置此方法，fullScreen()方法自动为true)
-     *              .transparentBar()             //透明状态栏和导航栏，不写默认状态栏为透明色，导航栏为黑色（设置此方法，fullScreen()方法自动为true）
-     *              .statusBarColor(R.color.colorPrimary)     //状态栏颜色，不写默认透明色
-     *              .navigationBarColor(R.color.colorPrimary) //导航栏颜色，不写默认黑色
-     *              .barColor(R.color.colorPrimary)  //同时自定义状态栏和导航栏颜色，不写默认状态栏为透明色，导航栏为黑色
-     *              .statusBarAlpha(0.3f)  //状态栏透明度，不写默认0.0f
-     *              .navigationBarAlpha(0.4f)  //导航栏透明度，不写默认0.0F
-     *              .barAlpha(0.3f)  //状态栏和导航栏透明度，不写默认0.0f
-     *              .statusBarDarkFont(true)   //状态栏字体是深色，不写默认为亮色
-     *              .navigationBarDarkIcon(true) //导航栏图标是深色，不写默认为亮色
-     *              .autoDarkModeEnable(true) //自动状态栏字体和导航栏图标变色，必须指定状态栏颜色和导航栏颜色才可以自动变色哦
-     *              .autoStatusBarDarkModeEnable(true,0.2f) //自动状态栏字体变色，必须指定状态栏颜色才可以自动变色哦
-     *              .autoNavigationBarDarkModeEnable(true,0.2f) //自动导航栏图标变色，必须指定导航栏颜色才可以自动变色哦
-     *              .flymeOSStatusBarFontColor(R.color.btn3)  //修改flyme OS状态栏字体颜色
-     *              .fullScreen(true)      //有导航栏的情况下，activity全屏显示，也就是activity最下面被导航栏覆盖，不写默认非全屏
-     *              .hideBar(BarHide.FLAG_HIDE_BAR)  //隐藏状态栏或导航栏或两者，不写默认不隐藏
-     *              .addViewSupportTransformColor(toolbar)  //设置支持view变色，可以添加多个view，不指定颜色，默认和状态栏同色，还有两个重载方法
-     *              .titleBar(view)    //解决状态栏和布局重叠问题，任选其一
-     *              .titleBarMarginTop(view)     //解决状态栏和布局重叠问题，任选其一
-     *              .statusBarView(view)  //解决状态栏和布局重叠问题，任选其一
-     *              .fitsSystemWindows(true)    //解决状态栏和布局重叠问题，任选其一，默认为false，当为true时一定要指定statusBarColor()，不然状态栏为透明色，还有一些重载方法
-     *              .supportActionBar(true) //支持ActionBar使用
-     *              .statusBarColorTransform(R.color.orange)  //状态栏变色后的颜色
-     *              .navigationBarColorTransform(R.color.orange) //导航栏变色后的颜色
-     *              .barColorTransform(R.color.orange)  //状态栏和导航栏变色后的颜色
-     *              .removeSupportView(toolbar)  //移除指定view支持
-     *              .removeSupportAllView() //移除全部view支持
-     *              .navigationBarEnable(true)   //是否可以修改导航栏颜色，默认为true
-     *              .navigationBarWithKitkatEnable(true)  //是否可以修改安卓4.4和emui3.x手机导航栏颜色，默认为true
-     *              .navigationBarWithEMUI3Enable(true) //是否可以修改emui3.x手机导航栏颜色，默认为true
-     *              .keyboardEnable(true)  //解决软键盘与底部输入框冲突问题，默认为false，还有一个重载方法，可以指定软键盘mode
-     *              .keyboardMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE)  //单独指定软键盘模式
-     *              .setOnKeyboardListener(new OnKeyboardListener() {    //软键盘监听回调，keyboardEnable为true才会回调此方法
-     *                    @Override
-     *                    public void onKeyboardChange(boolean isPopup, int keyboardHeight) {
-     *                        LogUtils.e(isPopup);  //isPopup为true，软键盘弹出，为false，软键盘关闭
-     *                    }
-     *               })
-     *              .setOnNavigationBarListener(onNavigationBarListener) //导航栏显示隐藏监听，目前只支持华为和小米手机
-     *              .setOnBarListener(OnBarListener) //第一次调用和横竖屏切换都会触发，可以用来做刘海屏遮挡布局控件的问题
-     *              .addTag("tag")  //给以上设置的参数打标记
-     *              .getTag("tag")  //根据tag获得沉浸式参数
-     *              .reset()  //重置所以沉浸式参数
-     *              .init();  //必须调用方可应用以上所配置的参数
+     *             .transparentStatusBar()  // 透明状态栏，不写默认透明色
+     *             .transparentNavigationBar()  // 透明导航栏，不写默认黑色(设置此方法，fullScreen()方法自动为 true)
+     *             .transparentBar()             // 透明状态栏和导航栏，不写默认状态栏为透明色，导航栏为黑色（设置此方法，fullScreen()方法自动为 true）
+     *             .statusBarColor(R.color.colorPrimary)     // 状态栏颜色，不写默认透明色
+     *             .navigationBarColor(R.color.colorPrimary) // 导航栏颜色，不写默认黑色
+     *             .barColor(R.color.colorPrimary)  // 同时自定义状态栏和导航栏颜色，不写默认状态栏为透明色，导航栏为黑色
+     *             .statusBarAlpha(0.3f)  // 状态栏透明度，不写默认 0.0f
+     *             .navigationBarAlpha(0.4f)  // 导航栏透明度，不写默认 0.0F
+     *             .barAlpha(0.3f)  // 状态栏和导航栏透明度，不写默认 0.0f
+     *             .statusBarDarkFont(true)   // 状态栏字体是深色，不写默认为亮色
+     *             .navigationBarDarkIcon(true) // 导航栏图标是深色，不写默认为亮色
+     *             .autoDarkModeEnable(true) // 自动状态栏字体和导航栏图标变色，必须指定状态栏颜色和导航栏颜色才可以自动变色哦
+     *             .autoStatusBarDarkModeEnable(true,0.2f) // 自动状态栏字体变色，必须指定状态栏颜色才可以自动变色哦
+     *             .autoNavigationBarDarkModeEnable(true,0.2f) // 自动导航栏图标变色，必须指定导航栏颜色才可以自动变色哦
+     *             .flymeOSStatusBarFontColor(R.color.btn3)  // 修改 flyme OS 状态栏字体颜色
+     *             .fullScreen(true)      // 有导航栏的情况下，activity 全屏显示，也就是 activity 最下面被导航栏覆盖，不写默认非全屏
+     *             .hideBar(BarHide.FLAG_HIDE_BAR)  // 隐藏状态栏或导航栏或两者，不写默认不隐藏
+     *             .addViewSupportTransformColor(toolbar)  // 设置支持 view 变色，可以添加多个 view，不指定颜色，默认和状态栏同色，还有两个重载方法
+     *             .titleBar(view)    // 解决状态栏和布局重叠问题，任选其一
+     *             .titleBarMarginTop(view)     // 解决状态栏和布局重叠问题，任选其一
+     *             .statusBarView(view)  // 解决状态栏和布局重叠问题，任选其一
+     *             .fitsSystemWindows(true)    // 解决状态栏和布局重叠问题，任选其一，默认为 false，当为 true 时一定要指定 statusBarColor()，不然状态栏为透明色，还有一些重载方法
+     *             .supportActionBar(true) // 支持 ActionBar 使用
+     *             .statusBarColorTransform(R.color.orange)  // 状态栏变色后的颜色
+     *             .navigationBarColorTransform(R.color.orange) // 导航栏变色后的颜色
+     *             .barColorTransform(R.color.orange)  // 状态栏和导航栏变色后的颜色
+     *             .removeSupportView(toolbar)  // 移除指定 view 支持
+     *             .removeSupportAllView() // 移除全部 view 支持
+     *             .navigationBarEnable(true)   // 是否可以修改导航栏颜色，默认为 true
+     *             .navigationBarWithKitkatEnable(true)  // 是否可以修改安卓 4.4 和 emui3.x 手机导航栏颜色，默认为 true
+     *             .navigationBarWithEMUI3Enable(true) // 是否可以修改 emui3.x 手机导航栏颜色，默认为true
+     *             .keyboardEnable(true)  // 解决软键盘与底部输入框冲突问题，默认为 false，还有一个重载方法，可以指定软键盘 mode
+     *             .keyboardMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE)  // 单独指定软键盘模式
+     *             .setOnKeyboardListener(new OnKeyboardListener() {    // 软键盘监听回调，keyboardEnable 为 true 才会回调此方法
+     *                   @Override
+     *                   public void onKeyboardChange(boolean isPopup, int keyboardHeight) {
+     *                       LogUtils.e(isPopup);  // isPopup 为 true，软键盘弹出，为 false，软键盘关闭
+     *                   }
+     *              })
+     *             .setOnNavigationBarListener(onNavigationBarListener) // 导航栏显示隐藏监听
+     *             .setOnStatusBarListener(onStatusBarListener) // 状态栏显示隐藏监听
+     *             .setOnBarListener(OnBarListener) // 第一次调用、横竖屏切换、状态栏/导航栏显隐变化都会触发，可以用来获取 BarProperties 快照
+     *             .addTag("tag")  // 给以上设置的参数打标记
+     *             .getTag("tag")  // 根据 tag 获得沉浸式参数
+     *             .reset()  // 重置所以沉浸式参数
+     *             .init();  // 必须调用方可应用以上所配置的参数
      */
     override fun initImmersionBar(statusBarDark: Boolean, navigationBarDark: Boolean, navigationBarColor: Int) {
         super.initImmersionBar(statusBarDark, navigationBarDark, navigationBarColor)
@@ -462,8 +466,11 @@ abstract class BaseActivity<VDB : ViewDataBinding> : AppCompatActivity(), BaseIm
     // <editor-fold defaultstate="collapsed" desc="页面管理方法">
     /**
      * ViewModel 中定义无值事件（用 Unit 替代 Any）
-     * val reason by lazy { MutableLiveData<Unit>() } // 无值事件
-     * Unit 类型的 value 是 Unit 实例（非 null），会触发回调
+     *  val reason by lazy { MutableLiveData<Unit>() } // 无值事件
+     *  Unit 类型的 value 是 Unit 实例（非 null），会触发回调
+     * putIfAbsent：
+     * 1) Key 不存在：插入新值，返回 null。
+     * 2) Key 已存在：不覆盖，返回已存在的那个旧 Value 对象本身（绝对不是 null，也不是什么“空值”）
      */
     protected fun <T> MutableLiveData<T>?.observe(block: T.() -> Unit) {
         this ?: return
@@ -701,7 +708,7 @@ abstract class BaseActivity<VDB : ViewDataBinding> : AppCompatActivity(), BaseIm
     override fun showDialog(flag: Boolean, second: Long, block: () -> Unit) {
         loadingDialog.apply { setDialogCancelable(flag) }.show()
         if (second > 0) {
-            TimerBuilder.schedule(this, {
+            schedule({
                 hideDialog()
                 block.invoke()
             }, second)
